general clean up for doc
