+++ forc-plugins/forc-doc/src/doc/module.rs
@@ -1,157 +1,11 @@
-use crate::{
-    descriptor::Descriptor,
-    render::{
-        constant::INDEX_FILENAME, item::components::*, link::DocLink, util::format::docstring::*,
-        Renderable,
-    },
-    RenderPlan,
-};
+use crate::render::{constant::INDEX_FILENAME, util::format::docstring::create_preview};
 use anyhow::Result;
-use horrorshow::{box_html, RenderBox, Template};
-use std::{fmt::Write, option::Option, path::PathBuf};
-use sway_core::{
-    decl_engine::DeclEngine,
-    language::{
-        ty::{TyAstNodeContent, TyProgram, TySubmodule},
-        CallPath,
-    },
-};
-
-pub(crate) type Documentation = Vec<Document>;
-/// A finalized Document ready to be rendered. We want to retain all
-/// information including spans, fields on structs, variants on enums etc.
-#[derive(Clone, Debug)]
-pub(crate) struct Document {
-    pub(crate) module_info: ModuleInfo,
-    pub(crate) item_header: ItemHeader,
-    pub(crate) item_body: ItemBody,
-    pub(crate) raw_attributes: Option<String>,
-}
-impl Document {
-    /// Creates an HTML file name from the [Document].
-    pub(crate) fn html_filename(&self) -> String {
-        use sway_core::language::ty::TyDecl::StorageDecl;
-        let name = match &self.item_body.ty_decl {
-            StorageDecl { .. } => None,
-            _ => Some(self.item_header.item_name.as_str()),
-        };
-
-        Document::create_html_filename(self.item_body.ty_decl.doc_name(), name)
-    }
-    fn create_html_filename(ty: &str, name: Option<&str>) -> String {
-        match name {
-            Some(name) => format!("{ty}.{name}.html"),
-            None => {
-                format!("{ty}.html") // storage does not have an Ident
-            }
-        }
-    }
-    /// Generate link info used in navigation between docs.
-    pub(crate) fn link(&self) -> DocLink {
-        DocLink {
-            name: self.item_header.item_name.as_str().to_owned(),
-            module_info: self.module_info.clone(),
-            html_filename: self.html_filename(),
-            preview_opt: self.preview_opt(),
-        }
-    }
-    fn preview_opt(&self) -> Option<String> {
-        create_preview(self.raw_attributes.clone())
-    }
-    /// Gather [Documentation] from the [TyProgram].
-    pub(crate) fn from_ty_program(
-        decl_engine: &DeclEngine,
-        project_name: &str,
-        typed_program: &TyProgram,
-        no_deps: bool,
-        document_private_items: bool,
-    ) -> Result<Documentation> {
-        // the first module prefix will always be the project name
-        let mut docs: Documentation = Default::default();
-        for ast_node in &typed_program.root.all_nodes {
-            if let TyAstNodeContent::Declaration(ref decl) = ast_node.content {
-                let desc = Descriptor::from_typed_decl(
-                    decl_engine,
-                    decl,
-                    ModuleInfo::from_ty_module(vec![project_name.to_owned()], None),
-                    document_private_items,
-                )?;
-
-                if let Descriptor::Documentable(doc) = desc {
-                    docs.push(doc)
-                }
-            }
-        }
-
-        if !no_deps && !typed_program.root.submodules.is_empty() {
-            // this is the same process as before but for dependencies
-            for (_, ref typed_submodule) in &typed_program.root.submodules {
-                let attributes = (!typed_submodule.module.attributes.is_empty())
-                    .then(|| typed_submodule.module.attributes.to_html_string());
-                let module_prefix =
-                    ModuleInfo::from_ty_module(vec![project_name.to_owned()], attributes);
-                Document::from_ty_submodule(
-                    decl_engine,
-                    typed_submodule,
-                    &mut docs,
-                    &module_prefix,
-                    document_private_items,
-                )?;
-            }
-        }
-
-        Ok(docs)
-    }
-    fn from_ty_submodule(
-        decl_engine: &DeclEngine,
-        typed_submodule: &TySubmodule,
-        docs: &mut Documentation,
-        module_prefix: &ModuleInfo,
-        document_private_items: bool,
-    ) -> Result<()> {
-        let mut new_submodule_prefix = module_prefix.to_owned();
-        new_submodule_prefix
-            .module_prefixes
-            .push(typed_submodule.mod_name_span.as_str().to_owned());
-        for ast_node in &typed_submodule.module.all_nodes {
-            if let TyAstNodeContent::Declaration(ref decl) = ast_node.content {
-                let desc = Descriptor::from_typed_decl(
-                    decl_engine,
-                    decl,
-                    new_submodule_prefix.clone(),
-                    document_private_items,
-                )?;
-
-                if let Descriptor::Documentable(doc) = desc {
-                    docs.push(doc)
-                }
-            }
-        }
-        for (_, submodule) in &typed_submodule.module.submodules {
-            Document::from_ty_submodule(
-                decl_engine,
-                submodule,
-                docs,
-                &new_submodule_prefix,
-                document_private_items,
-            )?;
-        }
-
-        Ok(())
-    }
-}
-impl Renderable for Document {
-    fn render(self, render_plan: RenderPlan) -> Result<Box<dyn RenderBox>> {
-        let header = self.item_header.render(render_plan.clone())?;
-        let body = self.item_body.render(render_plan)?;
-        Ok(box_html! {
-            : header;
-            : body;
-        })
-    }
-}
+use horrorshow::{box_html, Template};
+use std::{fmt::Write, path::PathBuf};
+use sway_core::language::CallPath;
 
 pub(crate) type ModulePrefixes = Vec<String>;
+
 #[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq)]
 pub(crate) struct ModuleInfo {
     pub(crate) module_prefixes: ModulePrefixes,
@@ -331,35 +185,6 @@ impl ModuleInfo {
     pub(crate) fn preview_opt(&self) -> Option<String> {
         create_preview(self.attributes.clone())
     }
-}
-
-/// Create a docstring preview from raw html attributes.
-///
-/// Returns `None` if there are no attributes.
-fn create_preview(raw_attributes: Option<String>) -> Option<String> {
-    const MAX_PREVIEW_CHARS: usize = 100;
-    const CLOSING_PARAGRAPH_TAG: &str = "</p>";
-
-    raw_attributes.as_ref().map(|description| {
-        let preview = split_at_markdown_header(description);
-        if preview.chars().count() > MAX_PREVIEW_CHARS && preview.contains(CLOSING_PARAGRAPH_TAG) {
-            let closing_tag_index = preview
-                .find(CLOSING_PARAGRAPH_TAG)
-                .expect("closing tag out of range");
-            // We add 1 here to get the index of the char after the closing tag.
-            // This ensures we retain the closing tag and don't break the html.
-            let (preview, _) =
-                preview.split_at(closing_tag_index + CLOSING_PARAGRAPH_TAG.len() + 1);
-            if preview.chars().count() > MAX_PREVIEW_CHARS && preview.contains('\n') {
-                let newline_index = preview.find('\n').expect("new line char out of range");
-                preview.split_at(newline_index).0.to_string()
-            } else {
-                preview.to_string()
-            }
-        } else {
-            preview.to_string()
-        }
-    })
 }
 
 #[cfg(test)]
